Improved new user creation and early registration
Now, when a new user logs in for the first time, a bikeUser is created
and the user is redirected to the registration page. Upon successful
registration, they are redirected to the profile page.  When a user
attempts to visit the profile page, they are redirected to register if
their registration has expired.