Added support for different types of issues
There are now three panels for the checkout person to use to submit
issues. They are in an accordion setup, so only one is visible at a
time. The panels allow for submitting just a charge for a user, a charge
and a maintenance report simultaneously, or just a maintenance issue.